Compilable using x64 compiler (but not yet tested)

+++ gui/NumeReWindow.cpp
@@ -2163,7 +2163,7 @@ void NumeReWindow::openTable(NumeRe::Container<std::string> _stringTable, const 
     frame->SetSize(800,600);
     TableViewer* grid = new TableViewer(frame, wxID_ANY, frame->CreateStatusBar(3), nullptr, wxDefaultPosition, wxDefaultSize, wxWANTS_CHARS | wxBORDER_STATIC);
     grid->SetData(_stringTable, tableDisplayName, sIntName);
-    frame->SetSize(std::min(800u, grid->GetWidth()), std::max(std::min(600u, grid->GetHeight()+50), 300u));
+    frame->SetSize(std::min((size_t)800u, grid->GetWidth()), std::max(std::min((size_t)600u, grid->GetHeight()+50), (size_t)300u));
     frame->SetIcon(getStandardIcon());
     frame->Show();
     frame->SetFocus();
@@ -2189,7 +2189,7 @@ void NumeReWindow::openTable(NumeRe::Table _table, const std::string& tableDispl
     TablePanel* panel = new TablePanel(frame, wxID_ANY, frame->CreateStatusBar(3));
     panel->SetTerminal(m_terminal);
     panel->grid->SetData(_table, tableDisplayName, sIntName);
-    frame->SetSize(std::min(800u, panel->grid->GetWidth()+200), std::max(std::min(600u, panel->grid->GetHeight()+50), 300u));
+    frame->SetSize(std::min((size_t)800u, panel->grid->GetWidth()+200), std::max(std::min((size_t)600u, panel->grid->GetHeight()+50), (size_t)300u));
     frame->SetIcon(getStandardIcon());
     frame->Show();
     frame->SetFocus();
@@ -2213,7 +2213,7 @@ void NumeReWindow::editTable(NumeRe::Container<std::string> _stringTable, const 
     TableEditPanel* panel = new TableEditPanel(frame, wxID_ANY, frame->CreateStatusBar(3));
     panel->SetTerminal(m_terminal);
     panel->grid->SetData(_stringTable, tableDisplayName, "");
-    frame->SetSize(std::min(800u, panel->grid->GetWidth()), std::max(std::min(600u, panel->grid->GetHeight()+50), 300u));
+    frame->SetSize(std::min((size_t)800u, panel->grid->GetWidth()), std::max(std::min((size_t)600u, panel->grid->GetHeight()+50), (size_t)300u));
     frame->SetIcon(getStandardIcon());
     frame->Show();
     frame->SetFocus();
@@ -2237,7 +2237,7 @@ void NumeReWindow::editTable(NumeRe::Table _table, const std::string& tableDispl
     TableEditPanel* panel = new TableEditPanel(frame, wxID_ANY, frame->CreateStatusBar(3));
     panel->SetTerminal(m_terminal);
     panel->grid->SetData(_table, tableDisplayName, "");
-    frame->SetSize(std::min(800u, panel->grid->GetWidth()+200), std::max(std::min(600u, panel->grid->GetHeight()+50), 300u));
+    frame->SetSize(std::min((size_t)800u, panel->grid->GetWidth()+200), std::max(std::min((size_t)(size_t)600u, panel->grid->GetHeight()+50), (size_t)300u));
     frame->SetIcon(getStandardIcon());
     frame->Show();
     frame->SetFocus();
